move http2_connection to its own file
