+++ lib/grpc.dart
@@ -24,8 +24,9 @@ export 'src/auth/auth.dart'
 export 'src/client/call.dart' show ClientCall;
 export 'src/client/client.dart' show Client;
 export 'src/client/common.dart' show Response, ResponseStream, ResponseFuture;
-export 'src/client/connection.dart' show ConnectionState, Http2ClientConnection;
+export 'src/client/connection.dart' show ConnectionState;
 export 'src/client/http2_channel.dart' show ClientChannel;
+export 'src/client/http2_connection.dart' show ClientConnection, Http2ClientConnection;
 export 'src/client/method.dart' show ClientMethod;
 export 'src/client/options.dart'
     show
